requirements.txt
# requirements.txt

streamlit
google-generativeai
Pillow
pyautogui
pygetwindow
requests
mss
python-dotenv
MetaTrader5